Create a combined HUD for Ruth/Roth (ss-c)
Add support to handle multiple visible HUD panels, plus minimize the HUD
to a small bar.

This requires a HUD that has additional prim to provide a new root for
the rotation plus a prim to click on for the bar.

Also add hud-maker.lsl to build a new combo HUD in-world.  This has room
for 3 HUD panels, only the Alpha HUD is currently included, the skin
applier HUD will be next.